:hammer: DRY external editing
